feat: discount sum and generate file

+++ Sem1/E-Government/HomeWork1/backend/src/main/java/com/example/backend/dto/PaymentOrderDto.java
@@ -12,7 +12,7 @@ public class PaymentOrderDto {
     String payerName;
     String payerIban;
     PayerType payerType;
-    int sum;
+    long sum;
     String receiverName;
     String receiverIban;
 }
