+++ Sem1/E-Government/HomeWork1/backend/src/main/java/com/example/backend/model/PaymentOrder.java
@@ -25,7 +25,7 @@ public class PaymentOrder {
     private String payerIban;
     @Enumerated(EnumType.STRING)
     private PayerType payerType;
-    private int sum;
+    private long sum;
     private String receiverName;
     private String receiverIban;
 }
